requirements.txt

Flask
numpy
joblib
scikit-learn
python-dotenv
google-generativeai
requests
